requirements.txt
Flask==3.1.0
SQLAlchemy==2.0.36
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Bcrypt==1.0.1
Flask-WTF==1.2.2
Markdown==3.7
Pygments==2.19.1
python-dotenv==1.0.1
email-validator==2.2.0
